app/requirements.txt
python-dotenv
flask
matplotlib
psycopg2
click
flask_sqlalchemy
Flask-Migrate
pandas
xgboost
scikit-learn
imblearn
numpy
